fix: moved Exposure metrics from vuln management to security requirements
